Changes and documentation
