+++ compiler-plugin/src/main/kotlin/tech/mappie/resolving/primitives/PrimitiveBodyCollector.kt
@@ -7,9 +7,7 @@ import org.jetbrains.kotlin.ir.declarations.IrFunction
 import org.jetbrains.kotlin.ir.expressions.*
 import tech.mappie.BaseVisitor
 
-class PrimitiveBodyCollector(
-    private val declaration: IrFunction,
-) : BaseVisitor<IrExpression, Unit>() {
+class PrimitiveBodyCollector(private val declaration: IrFunction) : BaseVisitor<IrExpression, Unit>() {
 
     override fun visitBlockBody(body: IrBlockBody, data: Unit): IrExpression {
         return body.statements.single().accept(data)
